+++ docs/CHANGE_LOG.md
@@ -3,6 +3,7 @@
 ## Version 2.2.1:
 - Fixed all browser warnings about inputs / text areas having missing IDs / names.
 - Fixed a fault that caused the "onOptionsUpdated" custom trigger to be called multiple times when changes options on the Side Menu.
+- Added new setting "startOfWeekDay" (defaults to 0 for Monday), which states what day is considered the start of the week (Monday, Saturday, and Sunday are currently supported).
 - 
 
 <br>